+++ examples/data-sources/iosxr_l2vpn_xconnect_group/data-source.tf
@@ -1,3 +1,0 @@
-data "iosxr_l2vpn_xconnect_group" "example" {
-  group_name = "P2P"
-}