+++ examples/resources/iosxr_l2vpn/resource.tf
@@ -1,4 +1,9 @@
 resource "iosxr_l2vpn" "example" {
   description = "My L2VPN Description"
   router_id   = "1.2.3.4"
+  xconnect_groups = [
+    {
+      group_name = "P2P"
+    }
+  ]
 }