+++ examples/resources/iosxr_l2vpn_xconnect_group/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_l2vpn_xconnect_group.example "Cisco-IOS-XR-um-l2vpn-cfg:/l2vpn/xconnect/groups/group[group-name=P2P]"