+++ examples/resources/iosxr_l2vpn_xconnect_group/resource.tf
@@ -1,3 +1,0 @@
-resource "iosxr_l2vpn_xconnect_group" "example" {
-  group_name = "P2P"
-}