+++ internal/provider/data_source_iosxr_l2vpn_test.go
@@ -18,6 +18,7 @@ func TestAccDataSourceIosxrL2VPN(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("data.iosxr_l2vpn.test", "description", "My L2VPN Description"),
 					resource.TestCheckResourceAttr("data.iosxr_l2vpn.test", "router_id", "1.2.3.4"),
+					resource.TestCheckResourceAttr("data.iosxr_l2vpn.test", "xconnect_groups.0.group_name", "P2P"),
 				),
 			},
 		},
@@ -29,6 +30,9 @@ const testAccDataSourceIosxrL2VPNConfig = `
 resource "iosxr_l2vpn" "test" {
 	description = "My L2VPN Description"
 	router_id = "1.2.3.4"
+	xconnect_groups = [{
+		group_name = "P2P"
+	}]
 }
 
 data "iosxr_l2vpn" "test" {